+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.0.1",
     "@appbaseio/reactivesearch": "^3.1.1",
+    "@hurumap-ui/charts": "^0.3.0",
     "@hurumap-ui/content": "^0.0.1-alpha.5",
     "@hurumap-ui/core": "^0.3.0",
     "@material-ui/core": "^4.5.1",
