+++ src/components/ProfilePage/index.js
@@ -8,19 +8,18 @@ import { Grid } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 
 import InsightContainer from '@hurumap-ui/core/InsightContainer';
-import { useProfileLoader, ChartFactory } from '@hurumap-ui/core';
+import ChartFactory from '@hurumap-ui/charts/ChartFactory';
+import useProfileLoader from '@hurumap-ui/core/useProfileLoader';
 import { shareIndicator } from '@hurumap-ui/core/utils';
 
-import config from '../config';
+import config from '../../config';
 
-import Page from './Page';
-import ProfileDetail from './ProfileDetail';
-import ProfileSection, { ProfileSectionTitle } from './ProfileSection';
-import Section from './Section';
+import Page from '../Page';
+import ProfileDetail from '../ProfileDetail';
+import ProfileSection, { ProfileSectionTitle } from '../ProfileSection';
+import Section from '../Section';
 
-import { getSectionedCharts } from '../cms';
-
-import logo from '../assets/images/logo-white-all.png';
+import logo from '../../assets/images/logo-white-all.png';
 
 const MapIt = dynamic({
   ssr: false,
@@ -445,14 +444,6 @@ Profile.propTypes = {
       ).isRequired
     })
   ).isRequired
-};
-
-Profile.getInitialProps = async ({ query: { lang: queryLang } }) => {
-  const lang = queryLang || config.DEFAULT_LANG;
-  return {
-    sectionedCharts: await getSectionedCharts(lang),
-    language: lang
-  };
 };
 
 export default Profile;
