+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
-import { renderBlocks } from '@hurumap-ui/content';
+import renderBlocks from '@hurumap-ui/content/renderBlocks';
 import logo from '../assets/images/logo-white-all.png';
 
 import { getSitePage } from '../cms';
