+++ src/theme.js
@@ -1,4 +1,4 @@
-import { createTheme } from '@hurumap-ui/core';
+import createTheme from '@hurumap-ui/charts/styles/createTheme';
 
 const FONT_FAMILY_HEADING = '"Lora", serif';
 const FONT_FAMILY_TEXT = '"Muli", sans-serif';
